Made extended data rention feature optional.

+++ nsync.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
-GPG_KEY='nsa_backup_key'
+# If you want to use NSA's extended data retention feature, please specify your GPG key here.
+# GPG_KEY='nsa_backup_key'
 
 if [ -z "$1" ]; then
     echo "Please specify a file or folder to back up."
@@ -10,11 +11,16 @@ else
 fi
 
 tarfile="nsa-backup-$(date +%Y%m%d).tar.gz"
-filename=$tarfile.gpg
+filename=$tarfile
 
 echo "Preparing files..."
 tar -cvzf $tarfile $target
-gpg -e -r $GPG_KEY $tarfile && rm -f $tarfile
+
+if [ -n "$GPG_KEY" ]; then
+    echo "Processing for extended data retention..."
+    filename=$tarfile.gpg
+    gpg -e -r $GPG_KEY $tarfile && rm -f $tarfile
+fi
 
 echo "--------------------------------"
 echo "Sending backup to NSA's secure servers..."
